requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
blinker==1.8.2
cffi==2.0.0
click==8.1.7
cryptography==46.0.3
ecdsa==0.19.1
fastapi==0.109.0
Flask==3.0.3
h11==0.16.0
httptools==0.7.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.1
numpy==2.1.2
pandas==2.2.3
passlib==1.7.4
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.3
rsa==4.9.1
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.35.1
threadpoolctl==3.5.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.2
uvicorn==0.27.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.0.4
